reorganizing, after renaming of WordList to SpellingList
