+++ src/components/WordList/index.tsx
@@ -1,0 +1,3 @@
+/** @format */
+
+export { WordList as default } from "./WordList";
