+++ src/components/lists/SpellingListItems/index.tsx
@@ -1,0 +1,3 @@
+/** @format */
+
+export { default as SpellingList } from "./SpellingList";
