--- src/components/lists/WordListItems/index.tsx
@@ -1,1 +1,0 @@
-export { default as WordListItem } from "./WordListItem";
